+++ tests/python/relay/test_qnn_quantize.py
@@ -69,108 +69,5 @@ def test_quantize_op():
     test_float32_to_int8()
 
 
-def test_mxnet_quantization():
-
-    def quantize_test_driver(in_dtype, quant_args, out_dtype, in_data, verify_output_data):
-        shape = in_data.shape
-        input_data = relay.var("input_data", shape=shape, dtype=in_dtype)
-        min_range = quant_args['min_range']
-        max_range = quant_args['max_range']
-        quantized_output = relay.qnn.op.quantize_mxnet_min_max(input_data,
-                                                               min_range=min_range,
-                                                               max_range=max_range,
-                                                               out_dtype=out_dtype)
-        mod = relay.Function(relay.analysis.free_vars(quantized_output), quantized_output)
-        mod = relay.Module.from_expr(mod)
-        mod = relay.qnn.transform.CanonicalizeOps()(mod)
-        with relay.build_config(opt_level=3):
-            graph, lib, params = relay.build(mod, "llvm", params=None)
-            rt_mod = graph_runtime.create(graph, lib, ctx=tvm.cpu(0))
-            rt_mod.set_input(input_data=in_data)
-            rt_mod.set_input(**params)
-            rt_mod.run()
-            res = rt_mod.get_output(0).asnumpy()
-            np.testing.assert_equal(res, verify_output_data)
-            assert res.dtype == out_dtype
-
-    def test_float32_to_uint8():
-        data = np.array([-63.5, -63, -62.5, -62, -61.5, 62, 62.5, 63, 63.5, 64]) \
-            .astype('float32') \
-            .reshape((2, 5))
-        output = np.array([0, 1, 2, 3, 4, 251, 252, 253, 254, 255]) \
-            .astype('uint8') \
-            .reshape((2, 5))
-        quant_args = {"min_range": -63.5, "max_range": 64}
-        quantize_test_driver(in_dtype='float32', quant_args=quant_args, out_dtype='uint8',
-                             in_data=data, verify_output_data=output)
-
-    def test_float32_to_int8():
-        data = np.array([-63.5, -63, -62.5, -62, -61.5, 62, 62.5, 63, 63.5, 64]) \
-            .astype('float32') \
-            .reshape((2, 5))
-        output = np.array([-126, -125, -124, -123, -122, 123, 124, 125, 126, 127]) \
-            .astype('int8') \
-            .reshape((2, 5))
-        quant_args = {"min_range": -63.5, "max_range": 64}
-        quantize_test_driver(in_dtype='float32', quant_args=quant_args, out_dtype='int8',
-                             in_data=data, verify_output_data=output)
-
-    test_float32_to_uint8()
-    test_float32_to_int8()
-
-
-def test_mxnet_mkldnn_quantization():
-
-    def quantize_test_driver(in_dtype, quant_args, out_dtype, in_data, verify_output_data):
-        shape = in_data.shape
-        input_data = relay.var("input_data", shape=shape, dtype=in_dtype)
-        min_range = quant_args['min_range']
-        max_range = quant_args['max_range']
-        quantized_output = relay.qnn.op.quantize_mxnet_min_max(input_data,
-                                                               min_range=min_range,
-                                                               max_range=max_range,
-                                                               out_dtype=out_dtype,
-                                                               use_mkldnn=True)
-        mod = relay.Function(relay.analysis.free_vars(quantized_output), quantized_output)
-        mod = relay.Module.from_expr(mod)
-        mod = relay.qnn.transform.CanonicalizeOps()(mod)
-        with relay.build_config(opt_level=3):
-            graph, lib, params = relay.build(mod, "llvm", params=None)
-            rt_mod = graph_runtime.create(graph, lib, ctx=tvm.cpu(0))
-            rt_mod.set_input(input_data=in_data)
-            rt_mod.set_input(**params)
-            rt_mod.run()
-            res = rt_mod.get_output(0).asnumpy()
-            np.testing.assert_equal(res, verify_output_data)
-            assert res.dtype == out_dtype
-
-    def test_float32_to_uint8():
-        data = np.array([-63.5, -63, -62.5, -62, -61.5, 62, 62.5, 63, 63.5, 64]) \
-            .astype('float32') \
-            .reshape((2, 5))
-        output = np.array([0, 0, 0, 0, 0, 247, 249, 251, 253, 255]) \
-            .astype('uint8') \
-            .reshape((2, 5))
-        quant_args = {"min_range": -63.5, "max_range": 64}
-        quantize_test_driver(in_dtype='float32', quant_args=quant_args, out_dtype='uint8',
-                             in_data=data, verify_output_data=output)
-
-    def test_float32_to_int8():
-        data = np.array([-63.5, -63, -62.5, -62, -61.5, 62, 62.5, 63, 63.5, 64]) \
-            .astype('float32') \
-            .reshape((2, 5))
-        output = np.array([-126, -125, -124, -123, -122, 123, 124, 125, 126, 127]) \
-            .astype('int8') \
-            .reshape((2, 5))
-        quant_args = {"min_range": -63.5, "max_range": 64}
-        quantize_test_driver(in_dtype='float32', quant_args=quant_args, out_dtype='int8',
-                             in_data=data, verify_output_data=output)
-
-    test_float32_to_uint8()
-    test_float32_to_int8()
-
-
 if __name__ == "__main__":
     test_quantize_op()
-    test_mxnet_quantization()
-    test_mxnet_mkldnn_quantization()
